Drop node 0.10 from CI Separate main .js file onto several
