+++ src/compose.js
@@ -1,7 +1,8 @@
 import isFunction from './isFunction';
 import isObject from './isObject';
 import isComposable from './isComposable';
-import {merge, assign} from './merge';
+import merge from './merge';
+const assign = Object.assign;
 
 /**
  * Creates new factory instance.
