+++ src/extract-functions.js
@@ -1,0 +1,11 @@
+import isFunction from './isFunction';
+
+export default function (...args) {
+  let result = [];
+  for (let i = 0; i < args.length; i++) {
+    const arg = args[i];
+    if (isFunction(arg)) result.push(arg);
+    else if (Array.isArray(arg)) result = result.concat(arg.filter(isFunction));
+  }
+  return result.length === 0 ? undefined : result;
+}
