+++ src/standardise-descriptor.js
@@ -1,0 +1,66 @@
+import isObject from './isObject';
+import extractFunctions from './extract-functions';
+import merge from './merge';
+
+const assign = Object.assign;
+
+export default function ({
+  methods,
+
+  properties,
+  props,
+  refs,
+
+  initializers,
+  init,
+
+  deepProperties,
+  deepProps,
+
+  propertyDescriptors,
+
+  staticProperties,
+  statics,
+
+  staticDeepProperties,
+  deepStatics,
+
+  staticPropertyDescriptors,
+
+  configuration,
+  conf,
+
+  deepConfiguration,
+  deepConf
+} = {}) {
+  const p = isObject(props) || isObject(refs) || isObject(properties) ?
+    assign({}, props, refs, properties) : undefined;
+
+  let dp = isObject(deepProps) ? merge({}, deepProps) : undefined;
+  dp = isObject(deepProperties) ? merge(dp, deepProperties) : dp;
+
+  const sp = isObject(statics) || isObject(staticProperties) ?
+    assign({}, statics, staticProperties) : undefined;
+
+  let dsp = isObject(deepStatics) ? merge({}, deepStatics) : undefined;
+  dsp = isObject(staticDeepProperties) ? merge(dsp, staticDeepProperties) : dsp;
+
+  const c = isObject(conf) || isObject(configuration) ?
+    assign({}, conf, configuration) : undefined;
+
+  let dc = isObject(deepConf) ? merge({}, deepConf) : undefined;
+  dc = isObject(deepConfiguration) ? merge(dc, deepConfiguration) : dc;
+
+  return {
+    methods,
+    properties: p,
+    initializers: extractFunctions(init, initializers),
+    deepProperties: dp,
+    staticProperties: sp,
+    staticDeepProperties: dsp,
+    propertyDescriptors,
+    staticPropertyDescriptors,
+    configuration: c,
+    deepConfiguration: dc
+  };
+}
